+++ eslint.config.mjs
@@ -60,33 +60,90 @@ export default antfu(
   },
   {
     files: [
+      '**/no-mixed-spaces-and-tabs/README.md',
+    ],
+    name: 'local/no-mixed-spaces-and-tabs/readme',
+    rules: {
+      'style/no-mixed-spaces-and-tabs': 'off',
+    },
+  },
+  {
+    files: [
+      '**/no-trailing-spaces/README.md',
+    ],
+    name: 'local/no-trailing-spaces/readme',
+    rules: {
+      'format/prettier': 'off',
+      'style/no-trailing-spaces': 'off',
+    },
+  },
+  {
+    files: [
       'packages/eslint-plugin/{rules,utils}/**/*.ts',
       'packages/shared/utils/**/*.ts',
     ],
     ignores: ['**/*.test.ts'],
-    name: 'local/restricted-imports',
+    name: 'local/restrict-types',
     rules: {
-      'no-restricted-imports': ['error', {
+      'ts/no-restricted-imports': ['error', {
         patterns: [
           {
-            group: ['@typescript-eslint/utils', '@typescript-eslint/utils/*'],
+            group: ['@typescript-eslint/utils', '@typescript-eslint/utils/*', '@typescript-eslint/types'],
             importNames: [
               'TSESTree',
-              'RuleFunction',
-              'RuleListener',
-              'SourceCode',
-              'RuleFixer',
-              'ReportFixFunction',
-              'RuleContext',
+              'TSESLint',
+              'JSONSchema',
+              'RuleWithMetaAndName',
               'EcmaVersion',
               'ReportDescriptor',
+              'ReportFixFunction',
+              'RuleContext',
+              'RuleFixer',
+              'RuleFunction',
+              'RuleListener',
+              'RuleModule',
               'Scope',
+              'SourceCode',
             ],
             message: 'Import from "#types" instead',
           },
           {
             group: ['#test', '#test/*'],
             message: 'Should not import test utilities',
+          },
+        ],
+      }],
+      'ts/no-restricted-types': ['error', {
+        types: {
+          'Tree.Node': {
+            fixWith: 'ASTNode',
+            message: 'Import ASTNode from "#types" instead.',
+          },
+          'Tree.Token': {
+            fixWith: 'Token',
+            message: 'Import Token from "#types" instead.',
+          },
+        },
+      }],
+    },
+  },
+  {
+    files: [
+      'packages/eslint-plugin/rules/**/*.ts',
+      'packages/shared/types/**/*.ts',
+    ],
+    name: 'local/restrict-utils',
+    rules: {
+      'ts/no-restricted-imports': ['error', {
+        patterns: [
+          {
+            group: ['@typescript-eslint/utils', '@typescript-eslint/types'],
+            importNames: ['AST_NODE_TYPES', 'AST_TOKEN_TYPES', 'ASTUtils'],
+            message: 'Import from "#utils/ast" instead',
+          },
+          {
+            group: ['@typescript-eslint/utils/ast-utils'],
+            message: 'Import from "#utils/ast" instead',
           },
         ],
       }],