+++ packages/eslint-plugin/rules/jsx-curly-spacing/jsx-curly-spacing.ts
@@ -9,8 +9,9 @@
  * @author Erik Wendel
  */
 
-import type { ASTNode, RuleFixer, Token, Tree } from '#types'
+import type { ASTNode, RuleFixer, Token } from '#types'
 import type { BasicConfig, MessageIds, RuleOptions } from './types'
+import { isTokenOnSameLine } from '#utils/ast'
 import { createRule } from '#utils/create-rule'
 
 const SPACING = {
@@ -152,16 +153,6 @@ export default createRule<RuleOptions, MessageIds>({
     const childrenConfig = children ? normalizeConfig(children, defaultConfig, true) : null
 
     /**
-     * Determines whether two adjacent tokens have a newline between them.
-     * @param - The left token object.
-     * @param right - The right token object.
-     * @returns Whether or not there is a newline between the tokens.
-     */
-    function isMultiline(left: Tree.Token, right: Tree.Token) {
-      return left.loc.end.line !== right.loc.start.line
-    }
-
-    /**
      * Trims text of whitespace between two ranges
      * @param fixer - the eslint fixer object
      * @param fromLoc - the start location
@@ -191,7 +182,7 @@ export default createRule<RuleOptions, MessageIds>({
      * @param token - The token to use for the report.
      * @param spacing
      */
-    function reportNoBeginningNewline(node: ASTNode, token: Tree.Token, spacing: string) {
+    function reportNoBeginningNewline(node: ASTNode, token: Token, spacing: string) {
       context.report({
         node,
         loc: token.loc.start,
@@ -364,23 +355,23 @@ export default createRule<RuleOptions, MessageIds>({
       if (spacing === SPACING.always) {
         if (!sourceCode.isSpaceBetween(first, second))
           reportRequiredBeginningSpace(node, first)
-        else if (!config.allowMultiline && isMultiline(first, second))
+        else if (!config.allowMultiline && !isTokenOnSameLine(first, second))
           reportNoBeginningNewline(node, first, spacing)
 
         if (!sourceCode.isSpaceBetween(penultimate, last))
           reportRequiredEndingSpace(node, last)
-        else if (!config.allowMultiline && isMultiline(penultimate, last))
+        else if (!config.allowMultiline && !isTokenOnSameLine(penultimate, last))
           reportNoEndingNewline(node, last, spacing)
       }
       else if (spacing === SPACING.never) {
-        if (isMultiline(first, second)) {
+        if (!isTokenOnSameLine(first, second)) {
           if (!config.allowMultiline)
             reportNoBeginningNewline(node, first, spacing)
         }
         else if (sourceCode.isSpaceBetween(first, second)) {
           reportNoBeginningSpace(node, first)
         }
-        if (isMultiline(penultimate, last)) {
+        if (!isTokenOnSameLine(penultimate, last)) {
           if (!config.allowMultiline)
             reportNoEndingNewline(node, last, spacing)
         }