+++ packages/eslint-plugin/rules/keyword-spacing/keyword-spacing._js_.test.ts
@@ -2136,6 +2136,65 @@ run<RuleOptions, MessageIds>({
     },
 
     // ----------------------------------------------------------------------
+    // using
+    // ----------------------------------------------------------------------
+
+    {
+      code: '{}using a = b',
+      output: '{} using a = b',
+      parserOptions: { ecmaVersion: 2026 },
+      errors: expectedBefore('using'),
+    },
+    {
+      code: '{} using a = b',
+      output: '{}using a = b',
+      options: [NEITHER],
+      parserOptions: { ecmaVersion: 2026 },
+      errors: unexpectedBefore('using'),
+    },
+    {
+      code: '{}using a = b',
+      output: '{} using a = b',
+      options: [override('using', BOTH)],
+      parserOptions: { ecmaVersion: 2026 },
+      errors: expectedBefore('using'),
+    },
+    {
+      code: '{} using a = b',
+      output: '{}using a = b',
+      options: [override('using', NEITHER)],
+      parserOptions: { ecmaVersion: 2026 },
+      errors: unexpectedBefore('using'),
+    },
+    {
+      code: '{}await using a = b',
+      output: '{} await using a = b',
+      parserOptions: { ecmaVersion: 2026 },
+      errors: expectedBefore('await'),
+    },
+    {
+      code: '{} await using a = b',
+      output: '{}await using a = b',
+      options: [NEITHER],
+      parserOptions: { ecmaVersion: 2026 },
+      errors: unexpectedBefore('await'),
+    },
+    {
+      code: '{}await using a = b',
+      output: '{} await using a = b',
+      options: [override('await', BOTH)],
+      parserOptions: { ecmaVersion: 2026 },
+      errors: expectedBefore('await'),
+    },
+    {
+      code: '{} await using a = b',
+      output: '{}await using a = b',
+      options: [override('await', NEITHER)],
+      parserOptions: { ecmaVersion: 2026 },
+      errors: unexpectedBefore('await'),
+    },
+
+    // ----------------------------------------------------------------------
     // continue
     // ----------------------------------------------------------------------