+++ packages/eslint-plugin/rules/multiline-ternary/multiline-ternary.ts
@@ -4,7 +4,7 @@
  */
 
 import type { MessageIds, RuleOptions } from './types'
-import { isNotClosingParenToken, isTokenOnSameLine } from '#utils/ast'
+import { isNotClosingParenToken, isSingleLine, isTokenOnSameLine } from '#utils/ast'
 import { createRule } from '#utils/create-rule'
 
 export default createRule<RuleOptions, MessageIds>({
@@ -128,7 +128,7 @@ export default createRule<RuleOptions, MessageIds>({
           }
         }
         else {
-          if (allowSingleLine && node.loc.start.line === node.loc.end.line)
+          if (allowSingleLine && isSingleLine(node))
             return
 
           if (areTestAndConsequentOnSameLine) {