+++ packages/eslint-plugin/rules/no-extra-parens/no-extra-parens.ts
@@ -3,23 +3,27 @@
 import type { ASTNode, Token, Tree } from '#types'
 import type { MessageIds, RuleOptions } from './types'
 import {
+  AST_NODE_TYPES,
   canTokensBeAdjacent,
   getPrecedence,
   getStaticPropertyName,
   isClosingParenToken,
   isDecimalInteger,
+  isKeywordToken,
   isMixedLogicalAndCoalesceExpressions,
   isNotClosingParenToken,
   isNotOpeningParenToken,
   isOpeningBraceToken,
   isOpeningBracketToken,
+  isOpeningParenToken,
   isParenthesized as isParenthesizedRaw,
+  isSingleLine,
+  isTokenOnSameLine,
   isTopLevelExpressionStatement,
+  isTypeAssertion,
   skipChainExpression,
 } from '#utils/ast'
 import { createRule } from '#utils/create-rule'
-import { AST_NODE_TYPES } from '@typescript-eslint/utils'
-import { isOpeningParenToken, isTypeAssertion } from '@typescript-eslint/utils/ast-utils'
 
 export default createRule<RuleOptions, MessageIds>({
   name: 'no-extra-parens',
@@ -276,8 +280,6 @@ export default createRule<RuleOptions, MessageIds>({
      */
     function ruleApplies(node: ASTNode) {
       if (node.type === 'JSXElement' || node.type === 'JSXFragment') {
-        const isSingleLine = node.loc.start.line === node.loc.end.line
-
         switch (IGNORE_JSX) {
           // Exclude this JSX element from linting
           case 'all':
@@ -285,11 +287,11 @@ export default createRule<RuleOptions, MessageIds>({
 
           // Exclude this JSX element if it is multi-line element
           case 'multi-line':
-            return isSingleLine
+            return isSingleLine(node)
 
           // Exclude this JSX element if it is single-line element
           case 'single-line':
-            return !isSingleLine
+            return !isSingleLine(node)
 
           // Nothing special to be done for JSX elements
           case 'none':
@@ -315,7 +317,7 @@ export default createRule<RuleOptions, MessageIds>({
      * @private
      */
     function isParenthesised(node: ASTNode) {
-      return isParenthesizedRaw(node, sourceCode, 1)
+      return isParenthesizedRaw(1, node, sourceCode)
     }
 
     /**
@@ -325,7 +327,7 @@ export default createRule<RuleOptions, MessageIds>({
      * @private
      */
     function isParenthesisedTwice(node: ASTNode) {
-      return isParenthesizedRaw(node, sourceCode, 2)
+      return isParenthesizedRaw(2, node, sourceCode)
     }
 
     /**
@@ -504,7 +506,7 @@ export default createRule<RuleOptions, MessageIds>({
      * @private
      */
     function hasExcessParensNoLineTerminator(token: Token, node: ASTNode) {
-      if (token.loc.end.line === node.loc.start.line)
+      if (isTokenOnSameLine(token, node))
         return hasExcessParens(node)
 
       return hasDoubleExcessParens(node)
@@ -865,7 +867,7 @@ export default createRule<RuleOptions, MessageIds>({
         isOpeningParenToken(firstToken)
         && (
           isOpeningBraceToken(secondToken)
-          || secondToken.type === 'Keyword' && (
+          || isKeywordToken(secondToken) && (
             secondToken.value === 'function'
             || secondToken.value === 'class'
             || secondToken.value === 'let'
@@ -875,7 +877,7 @@ export default createRule<RuleOptions, MessageIds>({
               || tokenAfterClosingParens.type === 'Identifier'
             )
           )
-          || secondToken && secondToken.type === 'Identifier' && secondToken.value === 'async' && thirdToken && thirdToken.type === 'Keyword' && thirdToken.value === 'function'
+          || secondToken && secondToken.type === 'Identifier' && secondToken.value === 'async' && isKeywordToken(thirdToken) && thirdToken.value === 'function'
         )
       ) {
         tokensToIgnore.add(secondToken)
@@ -1422,7 +1424,7 @@ export default createRule<RuleOptions, MessageIds>({
           const { argument } = node
           const operatorToken = sourceCode.getLastToken(node)!
 
-          if (argument.loc.end.line === operatorToken.loc.start.line) {
+          if (isTokenOnSameLine(argument, operatorToken)) {
             checkArgumentWithPrecedence(node)
           }
           else {